remove semaphores for better performance
break up trading server class into partial classes + update asynchronous API
