+++ Orderbook/OrderbookCS/TradingOrderbookMatching.cs
@@ -12,6 +12,7 @@ namespace TradingServer.OrderbookCS
             int type = (int) order.OrderType;
             Trades result = new Trades();
 
+            lock (_ordersLock)
             {
                 if (ImmediateHandleTypes.Contains(type))
                 {
