+++ TradingServer/TradingServerServiceProvider.cs
@@ -2,6 +2,6 @@ namespace TradingServer.Core
 {
     internal static class TradingServerServiceProvider 
     {
-        public static IServiceProvider? serviceProvider { get; set; }
+        public static IServiceProvider? ServiceProvider { get; set; }
     }
 }
